Combine libraries into single helpers file
